Virtual Visualizer version: 1.0
There are lots of devices for blind and virtually impaired people, but very few of them have multiple options at a suitable price. So the main target of my project was to develop a device that will help the blind and virtually impaired people. This device is a multi functional low cost device. It is an image processing base object detection project, uses Raspberry Pi 2 with ultrasonic sensor and USB camera. It uses image processing with ultrasonic sensor for detecting object as well as the size and shape of the objects using blob. We have implemented our project both in software side and hardware side.

This can be modified and designed for security purpose also.